I have V(down,up) cycles working with up>0.  (In a new way!)
The Graeser & Kornhuber description of V(1,1) cycles does not
really make sense.  Instead of splitting the obstacles in two,
one only needs to recognize that on the way down each correction
should be in K^{j-1} while on the way up we first update using
the prolonged correction, so then we can smooth in K^0 + ... + K^j.

That is, on the way down the obstacle for the correction is phi^j
but on the way up the obstacle is chi^j.

It is simple and it works.  On the two classical obstacle exact
solutions it looks like the best performance, measured in WU, is
with V(0,1) cycles, then V(1,0), and then V(1,1).  That is,
up-smoothing is a bit more powerful than down-smoothing, which
makes sense because the corrections are less constrained.

Added a -down 0 -up 1 regression test.

Also note that the "good" and "heuristic" decomposition figures
by class VisObstacle() no longer care about "up".