Good SIA smoother, sans regularization, if ice-free handled correctly.
The key logic is to test the pointwise Jacobian and if it gives
zero then don't move unless the source term pushes upward:

  if Jac[p] == 0.0:
     d = self.daccumulation if ell[p] > 0.0 else 0.0

Here PNGSSIA.daccumulation is set to 10 m.  This logic seems to get
the ice-free places unstuck if they need to be.  Otherwise the
smoother uses no regularization and it never divides by zero.

In particular, note PNGSSIA.eps is zero.  This regularization is only
needed for the theoretical concern of well-posedness.

Also: Clean up printing of Jac==0 places under -showsingular.
      Clean up comments etc.